+++ R/calculateExpectedRT.R
@@ -1,7 +1,7 @@
 #############################################################################
 #' Calculate Cumulants Lognormal Response Time Distribution
 #'
-#' These functions have been deprecated. See \code{\link{get_mean_3PLN}} or \code{\link{get_var_3PLN}} instead.
+#' These functions have been deprecated. See \code{\link{getMean3PLN}} or \code{\link{getVar3PLN}} instead.
 #'
 #'@param lambda Vector of time intensity parameters.
 #'@param phi [optional] Vector of speed sensitivity parameters.
@@ -12,12 +12,12 @@
 #' @describeIn get_mean_3PLN Calculate mean 3PLN
 #'@export
 calculateExpectedRT <- function(lambda, phi, zeta, sdEpsi) {
-  stop("This function is deprecated. See '?get_mean_3PLN' for more details.")
+  stop("This function is deprecated. See '?getMean3PLN' for more details.")
 }
 
 #' @describeIn get_mean_3PLN Calculate mean 2PLN
 #'@export
 calculateExpectedRTvar <- function(lambda, phi, zeta, sdEpsi) {
-  stop("This function is deprecated. See '?get_var_3PLN' for more details.")
+  stop("This function is deprecated. See '?getVar3PLN' for more details.")
 }